Move LinkTracker to its own module
